Add react-router & more cleanup

+++ server/src/graphql/typeDefs.ts
@@ -10,7 +10,7 @@ export const typeDefs = gql`
     numOfGuests: Int!
     numOfBeds: Int!
     numOfBaths: Int!
-    rating: Int!
+    rating: Float!
   }
 
   type Query {
